# api/factory_scheduling_openapi.yaml
openapi: 3.0.0
info:
  title: 工厂调度系统API
  description: 工厂调度系统的REST API接口文档，基于OptaPlanner实现的高级排程系统
  version: 1.0.0
  contact:
    name: 系统开发团队
    email: support@example.com
servers:
  - url: http://localhost:8080
    description: 开发环境服务器
paths:
...
  /api/machines:
    get:
      summary: 获取所有工作中心
      description: 查询系统中所有可用的工作中心（设备/机器）列表。
      tags:
        - 设备管理
      responses:
        '200':
          description: 获取工作中心列表成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWorkCenterList'
    post:
      summary: 创建工作中心
      description: 批量创建新的工作中心（设备/机器）记录。
      tags:
        - 设备管理
      requestBody:
        description: 工作中心对象列表
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/WorkCenter'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWorkCenterList'
  /api/machines/{id}:
    get:
      summary: 根据ID获取工作中心
      description: 根据指定的ID查询特定的工作中心（设备/机器）信息。
      tags:
        - 设备管理
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 工作中心ID
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWorkCenter'
        '404':
          description: 未找到指定ID的工作中心
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
    put:
      summary: 更新工作中心
      description: 根据指定的ID更新工作中心（设备/机器）的信息。
      tags:
        - 设备管理
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 工作中心ID
      requestBody:
        description: 工作中心对象，包含更新信息
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkCenter'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWorkCenter'
    delete:
      summary: 删除工作中心
      description: 根据指定的ID删除工作中心（设备/机器）记录。
      tags:
        - 设备管理
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 工作中心ID
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'

  # 订单管理相关接口
  /api/orders:
    get:
      summary: 获取所有订单
      description: 查询系统中所有的订单列表。
      tags:
        - 订单管理
      responses:
        '200':
          description: 获取订单列表成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderList'
    post:
      summary: 创建订单
      description: 批量创建新的订单记录。
      tags:
        - 订单管理
      requestBody:
        description: 订单对象列表
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderList'
  /api/orders/{id}:
    get:
      summary: 根据ID获取订单
      description: 根据指定的ID查询特定的订单信息。
      tags:
        - 订单管理
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 订单ID
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrder'
        '404':
          description: 未找到指定ID的订单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseError'
    delete:
      summary: 删除订单
      description: 根据指定的ID删除订单记录。
      tags:
        - 订单管理
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 订单ID
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'

  # 设备维护相关接口
...
components:
  schemas:
...
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          description: 响应状态码
        msg:
          type: string
          description: 响应消息
        reqId:
          type: string
          description: 请求ID
...
    ApiResponseError:
      type: object
      properties:
        code:
          type: integer
          description: 错误状态码
        msg:
          type: string
          description: 错误消息
        reqId:
          type: string
          description: 请求ID
...
    WorkCenter:
      type: object
      properties:
        id:
          type: string
          description: 工作中心ID
        workCenterCode:
          type: string
          description: 工作中心编码
        name:
          type: string
          description: 工作中心名称
        status:
          type: string
          description: 工作中心状态
    ApiResponseWorkCenter:
      type: object
      properties:
        code:
          type: integer
          description: 响应状态码
        msg:
          type: string
          description: 响应消息
        data:
          $ref: '#/components/schemas/WorkCenter'
        reqId:
          type: string
          description: 请求ID
    ApiResponseWorkCenterList:
      type: object
      properties:
        code:
          type: integer
          description: 响应状态码
        msg:
          type: string
          description: 响应消息
        data:
          type: array
          items:
            $ref: '#/components/schemas/WorkCenter'
        total:
          type: integer
          description: 数据总量
        reqId:
          type: string
          description: 请求ID

    # 订单相关
    Order:
      type: object
      properties:
        orderNo:
          type: string
          description: 订单编号（主键）
        erpStatus:
          type: string
          description: ERP系统状态
        orderStatus:
          type: string
          description: 订单状态
        planStartDate:
          type: string
          format: date
          description: 计划开始日期
        planEndDate:
          type: string
          format: date
          description: 计划结束日期
        factStartDate:
          type: string
          format: date-time
          description: 实际开始日期时间
        factEndDate:
          type: string
          format: date-time
          description: 实际结束日期时间
    ApiResponseOrder:
      type: object
      properties:
        code:
          type: integer
          description: 响应状态码
        msg:
          type: string
          description: 响应消息
        data:
          $ref: '#/components/schemas/Order'
        reqId:
          type: string
          description: 请求ID
    ApiResponseOrderList:
      type: object
      properties:
        code:
          type: integer
          description: 响应状态码
        msg:
          type: string
          description: 响应消息
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        total:
          type: integer
          description: 数据总量
        reqId:
          type: string
          description: 请求ID

    # 工序相关
...